добавлена фикстура для запуска выбранного браузера с выбранным языком.
