--- stepik_module_2/lesson2_3_step6.py
@@ -1,26 +1,0 @@
-import time
-import math
-
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-
-
-def calc(x):
-    return str(math.log(abs(12*math.sin(int(x)))))
-
-
-test_page = "http://suninjuly.github.io/redirect_accept.html"
-browser = webdriver.Chrome()
-browser.get(test_page)
-
-try:
-    browser.find_element(By.CSS_SELECTOR, "button.btn").click()
-    new_window = browser.window_handles[1]
-    browser.switch_to.window(new_window)
-    value = browser.find_element(By.ID, "input_value").text
-    res = calc(value)
-    browser.find_element(By.ID, "answer").send_keys(res)
-    browser.find_element(By.CSS_SELECTOR, "button.btn").click()
-finally:
-    time.sleep(10)
-    browser.quit()
