--- stepik_module_3/test_3_3_9_pytest_raises.py
@@ -1,28 +1,0 @@
-import pytest
-
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-from selenium.common.exceptions import NoSuchElementException
-
-
-def test_exception_1():
-    browser = webdriver.Chrome()
-
-    try:
-        browser.get("http://selenium1py.pythonanywhere.com/")
-        with pytest.raises(NoSuchElementException):
-            browser.find_element(By.CSS_SELECTOR, "button.btn")
-            pytest.fail("Не должно быть кнопки Отправить")
-    finally:
-        browser.quit()
-
-
-def test_exception_2():
-    browser = webdriver.Chrome()
-    try:
-        browser.get("http://selenium1py.pythonanywhere.com/")
-        with pytest.raises(NoSuchElementException):
-            browser.find_element(By.CSS_SELECTOR, "no_such_button.btn")
-            pytest.fail("Не должно быть кнопки Отправить")
-    finally:
-        browser.quit()
