--- stepik_module_3/test_abs.py
@@ -1,6 +1,0 @@
-def test_abs1():
-    assert abs(-42) == 42, "Should be absolute value of number"
-
-
-def test_abs2():
-    assert abs(-42) == -42, "Should be absolute value of number"
