--- stepik_module_3/test_abs_project.py
@@ -1,13 +1,0 @@
-import unittest
-
-
-class TestAbs(unittest.TestCase):
-    def test_abs1(self):
-        self.assertEqual(abs(-42), 42, "Should be absolute value of a number")
-
-    def test_abs2(self):
-        self.assertEqual(abs(-42), -42, "Should be absolute value of a number")
-
-
-if __name__ == '__main__':
-    unittest.main()
