--- stepik_module_3/test_fixture_1.py
@@ -1,44 +1,0 @@
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-
-
-link = "http://selenium1py.pythonanywhere.com/"
-
-class TestMainPage1():
-
-    @classmethod
-    def setup_class(cls):
-        print("\nstart browser for test suite...")
-        cls.browser = webdriver.Chrome()
-
-    @classmethod
-    def teardown_class(cls):
-        print("quit browser for test suite...")
-        cls.browser.quit()
-
-    def test_quest_should_see_login_link(self):
-        self.browser.get(link)
-        self.browser.find_element(By.ID, "login_link")
-
-    def test_quest_should_see_basket_link_on_main_page(self):
-        self.browser.get(link)
-        self.browser.find_element(By.CSS_SELECTOR, ".basket-mini .btn-group > a")
-
-
-class TestMainPage2():
-
-    def setup_method(self):
-        print("Start browser for test")
-        self.browser = webdriver.Chrome()
-
-    def teardown_method(self):
-        print("quit browser for test")
-        self.browser.quit()
-
-    def test_quest_should_see_login_link(self):
-        self.browser.get(link)
-        self.browser.find_element(By.ID, "login_link")
-
-    def test_quest_should_see_basket_link_on_main_page(self):
-        self.browser.get(link)
-        self.browser.find_element(By.CSS_SELECTOR, ".basket-mini .btn-group > a")
