--- stepik_module_3/test_fixture_2.py
@@ -1,24 +1,0 @@
-import pytest
-
-from selenium import webdriver
-from selenium.webdriver.common.by import By
-
-link = "http://selenium1py.pythonanywhere.com/"
-
-
-@pytest.fixture
-def browser():
-    print("\nstart browser for test...")
-    browser = webdriver.Chrome()
-    return browser
-
-
-class TestMainPAge1:
-
-    def test_quest_should_see_login_link(self, browser):
-        browser.get(link)
-        browser.find_element(By.ID, "login_link")
-
-    def test_guest_should_see_basket_link_on_the_main_page(self, browser):
-        browser.get(link)
-        browser.find_element_by_css_selector(".basket-mini .btn-group > a")
