--- stepik_module_3/test_rerun.py
@@ -1,11 +1,0 @@
-link = "http://selenium1py.pythonanywhere.com/"
-
-
-def test_guest_should_see_login_link_pass(browser):
-    browser.get(link)
-    browser.find_element_by_css_selector("#login_link")
-
-
-def test_guest_should_see_login_link_fail(browser):
-    browser.get(link)
-    browser.find_element_by_css_selector("#magic_link")
